feat: add PublicationWrapper
