+++ apps/web/src/components/Publication/QuotedPublication.tsx
@@ -1,5 +1,5 @@
+import PublicationWrapper from '@components/Shared/PublicationWrapper';
 import type { Publication } from '@lenster/lens';
-import { useRouter } from 'next/router';
 import type { FC } from 'react';
 
 import HiddenPublication from './HiddenPublication';
@@ -11,19 +11,10 @@ interface QuotedPublicationProps {
 }
 
 const QuotedPublication: FC<QuotedPublicationProps> = ({ publication }) => {
-  const { push } = useRouter();
-
   return (
-    <article
+    <PublicationWrapper
       className="cursor-pointer p-5 first:rounded-t-xl last:rounded-b-xl hover:bg-gray-100 dark:hover:bg-gray-900"
-      onClick={() => {
-        const selection = window.getSelection();
-        if (!selection || selection.toString().length === 0) {
-          push(`/posts/${publication?.id}`);
-        }
-      }}
-      data-testid={`publication-${publication.id}`}
-      aria-hidden="true"
+      publication={publication}
     >
       <PublicationHeader publication={publication} />
       <div className="ml-[53px]">
@@ -37,7 +28,7 @@ const QuotedPublication: FC<QuotedPublicationProps> = ({ publication }) => {
           />
         )}
       </div>
-    </article>
+    </PublicationWrapper>
   );
 };
